reservation page layout 완성
