+++ frontend/src/Routes/Reservation/SeatReservation.js
@@ -1,0 +1,6 @@
+import React from "react";
+import styled from "styled-components";
+
+const SeatReservation = () => {};
+
+export default SeatReservation;
