+++ frontend/src/Routes/RouteURL.js
@@ -1,7 +1,7 @@
 import Home from "Routes/Home";
 import Movies from "Routes/Movies";
 import MoviesInfo from "Routes/MoviesInfo";
-import Reservation from "Routes/Reservation";
+import Reservation from "Routes/Reservation/Reservation";
 import Theater from "Routes/Theater";
 import Event from "Routes/Event";
 import Store from "Routes/Store/Store";
